+++ packages/@lwc/babel-plugin-component/src/decorators/wire/transform.js
@@ -4,6 +4,7 @@
  * SPDX-License-Identifier: MIT
  * For full license text, see the LICENSE file in the repo root or https://opensource.org/licenses/MIT
  */
+const { isWireDecorator } = require('./shared');
 const { LWC_COMPONENT_PROPERTIES } = require('../../constants');
 
 const WIRE_PARAM_PREFIX = '$';
@@ -209,44 +210,52 @@ const scopedReferenceLookup = (scope) => (name) => {
     };
 };
 
-module.exports = function transform(t, wireDecorators) {
-    const wiredValues = wireDecorators.map(({ path }) => {
-        const [id, config] = path.get('expression.arguments');
+module.exports = function transform(t, decoratorMetas) {
+    const objectProperties = [];
+    const wireDecoratorMetas = decoratorMetas.filter(isWireDecorator);
+    if (wireDecoratorMetas.length) {
+        const wiredValues = wireDecoratorMetas.map(({ path }) => {
+            const [id, config] = path.get('expression.arguments');
 
-        const propertyName = path.parentPath.get('key.name').node;
-        const isClassMethod = path.parentPath.isClassMethod({
-            kind: 'method',
-        });
+            const propertyName = path.parentPath.get('key.name').node;
+            const isClassMethod = path.parentPath.isClassMethod({
+                kind: 'method',
+            });
 
-        const wiredValue = {
-            propertyName,
-            isClassMethod,
-        };
-
-        if (config) {
-            wiredValue.static = getWiredStatic(config);
-            wiredValue.params = getWiredParams(t, config);
-        }
-
-        const referenceLookup = scopedReferenceLookup(path.scope);
-        const isMemberExpression = id.isMemberExpression();
-        const isIdentifier = id.isIdentifier();
-
-        if (isIdentifier || isMemberExpression) {
-            const referenceName = isMemberExpression ? id.node.object.name : id.node.name;
-            const reference = referenceLookup(referenceName);
-            wiredValue.adapter = {
-                name: referenceName,
-                expression: t.cloneDeep(id.node),
-                reference: reference.type === 'module' ? reference.value : undefined,
+            const wiredValue = {
+                propertyName,
+                isClassMethod,
             };
-        }
 
-        return wiredValue;
-    });
+            if (config) {
+                wiredValue.static = getWiredStatic(config);
+                wiredValue.params = getWiredParams(t, config);
+            }
 
-    return t.objectProperty(
-        t.identifier(LWC_COMPONENT_PROPERTIES.WIRE),
-        buildWireConfigValue(t, wiredValues)
-    );
+            const referenceLookup = scopedReferenceLookup(path.scope);
+            const isMemberExpression = id.isMemberExpression();
+            const isIdentifier = id.isIdentifier();
+
+            if (isIdentifier || isMemberExpression) {
+                const referenceName = isMemberExpression ? id.node.object.name : id.node.name;
+                const reference = referenceLookup(referenceName);
+                wiredValue.adapter = {
+                    name: referenceName,
+                    expression: t.cloneDeep(id.node),
+                    reference: reference.type === 'module' ? reference.value : undefined,
+                };
+            }
+
+            return wiredValue;
+        });
+        objectProperties.push(
+            t.objectProperty(
+                t.identifier(LWC_COMPONENT_PROPERTIES.WIRE),
+                buildWireConfigValue(t, wiredValues)
+            )
+        );
+        wireDecoratorMetas.forEach(({ path }) => path.remove());
+    }
+
+    return objectProperties;
 };
